app/backend/back/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
reportlab==4.4.1
weasyprint==65.1
arabic-reshaper==3.0.0 
asn1crypto==1.5.1 
certifi==2025.4.26 
charset-normalizer==3.4.2 
cryptography==45.0.3 
html5lib==1.1 
idna==3.10 
lxml==5.4.0 
oscrypto==1.3.0 
pyHanko==0.29.0 
pyhanko-certvalidator==0.27.0 
pypdf==5.5.0 
python-bidi==0.6.6 
pyyaml==6.0.2 
requests==2.32.3 
six==1.17.0 
svglib==1.5.1 
tzlocal==5.3.1 
uritools==5.0.0 
urllib3==2.4.0 
xhtml2pdf==0.2.17
djangorestframework-simplejwt==5.5.0
pyjwt==1.7.1